requirements_telegram.txt
python-telegram-bot>=20.0
Pillow>=10.0.0
openai>=1.0.0
